requirements.txt
PySide6>=6.5.0
yt-dlp>=2023.7.6
pyinstaller>=5.13.0
requests>=2.31.0
pillow>=10.0.0